simplify by representing all status log header versions as a single enum
