+++ crates/skytem_logs/src/mbed_motor_control/status.rs
@@ -1,5 +1,4 @@
 pub mod entry;
 pub mod header_v1;
 pub mod header_v2;
-pub mod statuslog_v1;
-pub mod statuslog_v2;
+pub mod statuslog;
