+++ crates/skytem_logs/src/mbed_motor_control/status/statuslog.rs
@@ -1,0 +1,426 @@
+use byteorder::{LittleEndian, ReadBytesExt};
+use chrono::{DateTime, Utc};
+use log_if::prelude::*;
+use serde::{Deserialize, Serialize};
+use std::{
+    fmt, fs,
+    io::{self, Read},
+    path::Path,
+};
+
+use crate::{
+    mbed_motor_control::mbed_header::{
+        MbedMotorControlLogHeader, UniqueDescriptionData, SIZEOF_UNIQ_DESC,
+    },
+    parse_unique_description,
+};
+
+use super::{entry::StatusLogEntry, header_v1::StatusLogHeaderV1, header_v2::StatusLogHeaderV2};
+
+#[derive(Debug, Clone, PartialEq, Deserialize, Serialize)]
+pub enum StatusLogHeader {
+    V1(StatusLogHeaderV1),
+    V2(StatusLogHeaderV2),
+}
+
+impl fmt::Display for StatusLogHeader {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            Self::V1(h) => write!(f, "{h}"),
+            Self::V2(h) => write!(f, "{h}"),
+        }
+    }
+}
+
+impl StatusLogHeader {
+    const UNIQUE_DESCRIPTION: &'static str = "MBED-MOTOR-CONTROL-STATUS-LOG-2024";
+
+    fn from_reader(reader: &mut impl io::Read) -> io::Result<Self> {
+        let mut unique_description: UniqueDescriptionData = [0u8; SIZEOF_UNIQ_DESC];
+        reader.read_exact(&mut unique_description)?;
+        if parse_unique_description(&unique_description) != Self::UNIQUE_DESCRIPTION {
+            return Err(io::Error::new(
+                io::ErrorKind::InvalidData,
+                "Not an Mbed StatusLog",
+            ));
+        }
+        let version = reader.read_u16::<LittleEndian>()?;
+        let header = match version {
+            1 => Self::V1(StatusLogHeaderV1::from_reader_with_uniq_descr_version(
+                reader,
+                unique_description,
+                version,
+            )?),
+            2 => Self::V2(StatusLogHeaderV2::from_reader_with_uniq_descr_version(
+                reader,
+                unique_description,
+                version,
+            )?),
+            _ => {
+                return Err(io::Error::new(
+                    io::ErrorKind::InvalidData,
+                    "Unsupported version",
+                ))
+            }
+        };
+        Ok(header)
+    }
+}
+
+#[derive(Debug, Clone, PartialEq, Deserialize, Serialize)]
+pub struct StatusLog {
+    header: StatusLogHeader,
+    entries: Vec<StatusLogEntry>,
+    timestamp_ns: Vec<f64>,
+    labels: Vec<PlotLabels>,
+    all_plots_raw: Vec<RawPlot>,
+    startup_timestamp: DateTime<Utc>,
+}
+
+impl StatusLog {
+    const UNIQUE_DESCRIPTION: &'static str = "MBED-MOTOR-CONTROL-STATUS-LOG-2024";
+
+    /// Probes the buffer and check if it starts with [`Self::UNIQUE_DESCRIPTION`] and therefor contains a valid [`PidLog`]
+    pub fn is_buf_valid(content: &[u8]) -> bool {
+        if content.len() < SIZEOF_UNIQ_DESC + 2 {
+            return false;
+        }
+
+        let unique_description = &content[..SIZEOF_UNIQ_DESC];
+        parse_unique_description(unique_description) == Self::UNIQUE_DESCRIPTION
+    }
+
+    /// Checks if the file at the given path is a valid [`PidLog`] file
+    pub fn file_is_valid(path: &Path) -> bool {
+        let Ok(mut file) = fs::File::open(path) else {
+            return false;
+        };
+
+        let mut buffer = vec![0u8; SIZEOF_UNIQ_DESC + 2];
+        match file.read_exact(&mut buffer) {
+            Ok(_) => Self::is_buf_valid(&buffer),
+            Err(_) => false, // Return false if we can't read enough bytes
+        }
+    }
+
+    // helper function build all the plots that can be made from a statuslog
+    fn build_raw_plots(startup_timestamp_ns: f64, entries: &[StatusLogEntry]) -> Vec<RawPlot> {
+        let engine_temp_plot_raw = plot_points_from_log_entry(
+            entries,
+            |e| e.timestamp_ns() + startup_timestamp_ns,
+            |e| e.engine_temp as f64,
+        );
+        let fan_on_plot_raw = plot_points_from_log_entry(
+            entries,
+            |e| e.timestamp_ns() + startup_timestamp_ns,
+            |e| (e.fan_on as u8) as f64,
+        );
+        let vbat_plot_raw = plot_points_from_log_entry(
+            entries,
+            |e| e.timestamp_ns() + startup_timestamp_ns,
+            |e| e.vbat as f64,
+        );
+        let setpoint_plot_raw = plot_points_from_log_entry(
+            entries,
+            |e| e.timestamp_ns() + startup_timestamp_ns,
+            |e| e.setpoint as f64,
+        );
+        let motor_state_plot_raw = plot_points_from_log_entry(
+            entries,
+            |e| e.timestamp_ns() + startup_timestamp_ns,
+            |e| (e.motor_state as u8) as f64,
+        );
+        vec![
+            RawPlot::new(
+                "Engine Temp °C".into(),
+                engine_temp_plot_raw,
+                ExpectedPlotRange::OneToOneHundred,
+            ),
+            RawPlot::new(
+                "Fan On".into(),
+                fan_on_plot_raw,
+                ExpectedPlotRange::Percentage,
+            ),
+            RawPlot::new(
+                "Vbat [V]".into(),
+                vbat_plot_raw,
+                ExpectedPlotRange::OneToOneHundred,
+            ),
+            RawPlot::new(
+                "Setpoint".into(),
+                setpoint_plot_raw,
+                ExpectedPlotRange::Thousands,
+            ),
+            RawPlot::new(
+                "Motor State".into(),
+                motor_state_plot_raw,
+                ExpectedPlotRange::OneToOneHundred,
+            ),
+        ]
+    }
+}
+
+impl SkytemLog for StatusLog {
+    type Entry = StatusLogEntry;
+
+    fn from_reader<R: io::Read>(reader: &mut R) -> io::Result<Self> {
+        let header = StatusLogHeader::from_reader(reader)?;
+        let vec_of_entries: Vec<StatusLogEntry> = parse_to_vec(reader);
+        let startup_timestamp = match header {
+            StatusLogHeader::V1(h) => h.startup_timestamp(),
+            StatusLogHeader::V2(h) => h.startup_timestamp(),
+        }
+        .map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))?
+        .and_utc();
+        let startup_timestamp_ns = startup_timestamp
+            .timestamp_nanos_opt()
+            .expect("timestamp as nanoseconds out of range")
+            as f64;
+        let timestamp_ns: Vec<f64> = vec_of_entries
+            .iter()
+            .map(|e| startup_timestamp_ns + e.timestamp_ns())
+            .collect();
+
+        let timestamps_with_state_changes =
+            parse_timestamps_with_state_changes(&vec_of_entries, startup_timestamp_ns);
+        let labels = vec![PlotLabels::new(
+            timestamps_with_state_changes,
+            ExpectedPlotRange::OneToOneHundred,
+        )];
+
+        let all_plots_raw = Self::build_raw_plots(startup_timestamp_ns, &vec_of_entries);
+        // Iterate through the plots and make sure all the first timestamps match
+        if let Some(first_plot) = all_plots_raw.first() {
+            if let Some([first_timestamp, ..]) = first_plot.points().first() {
+                for p in &all_plots_raw {
+                    if let Some([current_first_timestamp, ..]) = p.points().first() {
+                        debug_assert_eq!(current_first_timestamp, first_timestamp, "First timestamp of plots are not equal, was an offset applied to some plots but not all?");
+                    }
+                }
+            }
+        }
+
+        Ok(Self {
+            header,
+            entries: vec_of_entries,
+            labels,
+            timestamp_ns,
+            all_plots_raw,
+            startup_timestamp,
+        })
+    }
+
+    fn entries(&self) -> &[Self::Entry] {
+        &self.entries
+    }
+}
+
+impl GitMetadata for StatusLog {
+    fn project_version(&self) -> Option<String> {
+        match &self.header {
+            StatusLogHeader::V1(h) => h.project_version(),
+            StatusLogHeader::V2(h) => h.project_version(),
+        }
+    }
+    fn git_short_sha(&self) -> Option<String> {
+        match &self.header {
+            StatusLogHeader::V1(h) => h.git_short_sha(),
+            StatusLogHeader::V2(h) => h.git_short_sha(),
+        }
+    }
+
+    fn git_branch(&self) -> Option<String> {
+        match &self.header {
+            StatusLogHeader::V1(h) => h.git_branch(),
+            StatusLogHeader::V2(h) => h.git_branch(),
+        }
+    }
+
+    fn git_repo_status(&self) -> Option<String> {
+        match &self.header {
+            StatusLogHeader::V1(h) => h.git_repo_status(),
+            StatusLogHeader::V2(h) => h.git_repo_status(),
+        }
+    }
+}
+
+impl Plotable for StatusLog {
+    fn raw_plots(&self) -> &[RawPlot] {
+        &self.all_plots_raw
+    }
+
+    fn first_timestamp(&self) -> chrono::DateTime<chrono::Utc> {
+        self.startup_timestamp
+    }
+
+    fn descriptive_name(&self) -> &str {
+        match self.header {
+            StatusLogHeader::V1(_) => "Mbed Status v1",
+            StatusLogHeader::V2(_) => "Mbed Status v2",
+        }
+    }
+
+    fn labels(&self) -> Option<&[PlotLabels]> {
+        Some(&self.labels)
+    }
+
+    fn metadata(&self) -> Option<Vec<(String, String)>> {
+        let metadata = vec![
+            (
+                "Project Version".to_owned(),
+                self.project_version().unwrap_or_else(|| "N/A".to_owned()),
+            ),
+            (
+                "Git Branch".to_owned(),
+                self.git_branch().unwrap_or_else(|| "N/A".to_owned()),
+            ),
+            (
+                "Git Repo Status".to_owned(),
+                self.git_repo_status().unwrap_or_else(|| "Clean".to_owned()),
+            ),
+            (
+                "Git Short SHA".to_owned(),
+                self.git_short_sha().unwrap_or_else(|| "N/A".to_owned()),
+            ),
+            (
+                "Startup Timestamp".to_owned(),
+                self.startup_timestamp.naive_utc().to_string(),
+            ),
+        ];
+
+        Some(metadata)
+    }
+}
+
+impl fmt::Display for StatusLog {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        writeln!(f, "Header: {}", self.header)?;
+        for e in &self.entries {
+            writeln!(f, "{e}")?;
+        }
+        Ok(())
+    }
+}
+
+fn parse_timestamps_with_state_changes(
+    entries: &[StatusLogEntry],
+    startup_timestamp_ns: f64,
+) -> Vec<([f64; 2], String)> {
+    let mut result = Vec::new();
+    let mut last_state = None;
+
+    for entry in entries {
+        // Check if the current state is different from the last recorded state
+        if last_state != Some(entry.motor_state) {
+            // apply negative offset if we're going to a state with a lower value
+            let offset = if last_state.is_some_and(|ls| ls as u8 > entry.motor_state as u8) {
+                -0.5
+            } else {
+                0.5
+            };
+            result.push((
+                [
+                    entry.timestamp_ns() + startup_timestamp_ns,
+                    (entry.motor_state as u8) as f64 + offset,
+                ],
+                entry.motor_state.to_string(),
+            ));
+            last_state = Some(entry.motor_state);
+        }
+    }
+    result
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use std::fs::{self, File};
+    use testresult::TestResult;
+
+    const TEST_DATA_V1: &str =
+        "../../test_data/mbed_motor_control/v1/20240926_121708/status_20240926_121708_00.bin";
+    const TEST_DATA_V2: &str =
+        "../../test_data/mbed_motor_control/v2/20240822_085220/status_20240822_085220_00.bin";
+
+    use crate::{mbed_motor_control::status::entry::MotorState, parse_and_display_log_entries};
+
+    #[test]
+    fn test_deserialize_v1() -> TestResult {
+        let data = fs::read(TEST_DATA_V1)?;
+        let status_log = StatusLog::from_reader(&mut data.as_slice())?;
+        eprintln!("{}", status_log.header);
+
+        let first_entry = status_log.entries().first().expect("Empty entries vec");
+        assert_eq!(first_entry.engine_temp, 64.394905);
+        assert!(!first_entry.fan_on);
+        assert_eq!(first_entry.vbat, 11.76342);
+        assert_eq!(first_entry.setpoint, 1800.0);
+        assert_eq!(first_entry.motor_state, MotorState::ECU_ON_WAIT_PUMP);
+        let second_entry = &status_log.entries[1];
+        assert_eq!(second_entry.engine_temp, 64.394905);
+        assert!(!second_entry.fan_on);
+        assert_eq!(second_entry.vbat, 11.718291);
+        assert_eq!(second_entry.setpoint, 1800.0);
+        assert_eq!(second_entry.motor_state, MotorState::ECU_ON_WAIT_PUMP);
+
+        let last_entry = status_log.entries().last().expect("Empty entries vec");
+        assert_eq!(last_entry.timestamp_ns(), 930624.0 * 1_000_000.0);
+        assert_eq!(last_entry.engine_temp, 77.31132);
+        assert!(!last_entry.fan_on);
+        assert_eq!(last_entry.vbat, 11.996582);
+        assert_eq!(last_entry.setpoint, 3600.0);
+        assert_eq!(last_entry.motor_state, MotorState::STANDBY_WAIT_FOR_CAP);
+        //eprintln!("{status_log}");
+        Ok(())
+    }
+
+    #[test]
+    fn test_parse_and_display_v1() -> TestResult {
+        let file = File::open(TEST_DATA_V1)?;
+        let mut reader = io::BufReader::new(file);
+        let header = StatusLogHeader::from_reader(&mut reader)?;
+        println!("{header}");
+        parse_and_display_log_entries::<StatusLogEntry, _>(&mut reader, Some(10));
+        Ok(())
+    }
+
+    #[test]
+    fn test_deserialize_v2() -> TestResult {
+        let data = fs::read(TEST_DATA_V2)?;
+        let status_log = StatusLog::from_reader(&mut data.as_slice())?;
+        eprintln!("{}", status_log.header);
+
+        let first_entry = status_log.entries().first().expect("Empty entries vec");
+        assert_eq!(first_entry.engine_temp, 4.770642);
+        assert!(!first_entry.fan_on);
+        assert_eq!(first_entry.vbat, 4.730086);
+        assert_eq!(first_entry.setpoint, 2500.0);
+        assert_eq!(first_entry.motor_state, MotorState::POWER_HOLD);
+        let second_entry = &status_log.entries[1];
+        assert_eq!(second_entry.engine_temp, 4.770642);
+        assert!(!second_entry.fan_on);
+        assert_eq!(second_entry.vbat, 4.75265);
+        assert_eq!(second_entry.setpoint, 2500.0);
+        assert_eq!(second_entry.motor_state, MotorState::POWER_HOLD);
+
+        let last_entry = status_log.entries().last().expect("Empty entries vec");
+        assert_eq!(last_entry.timestamp_ns(), 14846000000.0);
+        assert_eq!(last_entry.engine_temp, 4.770642);
+        assert!(!last_entry.fan_on);
+        assert_eq!(last_entry.vbat, 4.7075214);
+        assert_eq!(last_entry.setpoint, 2500.0);
+        assert_eq!(last_entry.motor_state, MotorState::POWER_HOLD);
+        //eprintln!("{status_log}");
+        Ok(())
+    }
+
+    #[test]
+    fn test_parse_and_display_v2() -> TestResult {
+        let file = File::open(TEST_DATA_V2)?;
+        let mut reader = io::BufReader::new(file);
+        let header = StatusLogHeaderV2::from_reader(&mut reader)?;
+        println!("{header}");
+        parse_and_display_log_entries::<StatusLogEntry, _>(&mut reader, Some(10));
+        Ok(())
+    }
+}
